Refactor of the entities identifiers complete.
Now the database entities have sequential PK IDs (private identifiers to improve data search and filtering) and UUIDs.

This change was necessary to achieve better performance on some complex queries and infinite scrolling techniques, as well as safeguarding for future scaling of the application in case one day the PK IDs needs to be changed (having UUIDs ensure we have a fallback in this specific case).